+++ enkf/common/ncutils.h
@@ -24,8 +24,8 @@ void ncu_writefield(char fname[], char varname[], int k, int ni, int nj, int nk,
 void ncu_writerow(char fname[], char varname[], int k, int j, float* v);
 void ncu_read3dfield(char fname[], char varname[], int ni, int nj, int nk, float* v);
 int ncu_getnD(char fname[], char varname[]);
-void ncu_readvarfloat(int ncid, int varid, int n, float v[]);
-void ncu_readvardouble(int ncid, int varid, int n, double v[]);
+void ncu_readvarfloat(int ncid, int varid, size_t n, float v[]);
+void ncu_readvardouble(int ncid, int varid, size_t n, double v[]);
 
 #define _NCUTILS_H
 #endif
